// src/components/types.ts
export type CSSPropertiesWithVars =
  | '--palette-ui-purple-900'
  | '--palette-ui-purple-800'
  | '--palette-ui-purple-700'
  | '--palette-ui-purple-600'
  | '--palette-ui-purple-500'
  | '--palette-ui-purple-400'
  | '--palette-ui-purple-300'
  | '--palette-ui-purple-200'
  | '--palette-ui-purple-100'
  | '--palette-ui-purple-50'
  | '--palette-ui-red-900'
  | '--palette-ui-red-800'
  | '--palette-ui-red-700'
  | '--palette-ui-red-600'
  | '--palette-ui-red-500'
  | '--palette-ui-red-400'
  | '--palette-ui-red-300'
  | '--palette-ui-red-200'
  | '--palette-ui-red-100'
  | '--palette-ui-red-25'
  | '--palette-ui-orange-900'
  | '--palette-ui-orange-800'
  | '--palette-ui-orange-700'
  | '--palette-ui-orange-600'
  | '--palette-ui-orange-500'
  | '--palette-ui-orange-400'
  | '--palette-ui-green-900'
  | '--palette-ui-green-800'
  | '--palette-ui-green-700'
  | '--palette-ui-green-600'
  | '--palette-ui-green-500'
  | '--palette-ui-green-400'
  | '--palette-ui-green-300'
  | '--palette-ui-green-200'
  | '--palette-ui-green-100'
  | '--palette-ui-black-900'
  | '--palette-ui-black-800'
  | '--palette-ui-black-700'
  | '--palette-ui-black-600'
  | '--palette-ui-black-500'
  | '--palette-ui-black-400'
  | '--palette-ui-black-300'
  | '--palette-ui-black-200'
  | '--palette-ui-black-100'
  | '--palette-ui-black-50'
  | '--palette-ui-black-25'
  | '--palette-ui-white-900';